backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
wsproto==1.2.0
python-multipart==0.0.20
pydantic==2.11.1
pydantic-settings==2.9.1
aiofiles==24.1.0
Pillow==11.2.1
oss2>=2.18,<3
